feat: add map caching for pokemons
