+++ src/shared/lib/cache.ts
@@ -1,0 +1,3 @@
+import type { PokemonData } from '../types/pokemon';
+
+export const pokemonCache = new Map<string, PokemonData>();
